test: Finish up NestableFolder integration test suite
